+++ src/data/persian/projects-data.js
@@ -1,0 +1,60 @@
+export const projectsData = [
+  {
+    img: "/img/home-decor-1.jpeg",
+    title: "مدرن",
+    tag: "پروژه #1",
+    description:
+      "همانطور که اوبر از طریق مقدار زیادی از آشفتگی مدیریت داخلی کار می کند.",
+    route: "/dashboard/profile",
+    members: [
+      { img: "/img/team-1.jpeg", name: "رومینا حدید" },
+      { img: "/img/team-2.jpeg", name: "رایان تامپسون" },
+      { img: "/img/team-3.jpeg", name: "جسیکا دو" },
+      { img: "/img/team-4.jpeg", name: "الکساندر اسمیت" },
+    ],
+  },
+  {
+    img: "/img/home-decor-2.jpeg",
+    title: "اسکاندیناوی",
+    tag: "پروژه #2",
+    description:
+      "موسیقی چیزی است که هر فرد نظر خاص خود را در مورد آن دارد.",
+    route: "/dashboard/profile",
+    members: [
+      { img: "/img/team-4.jpeg", name: "الکساندر اسمیت" },
+      { img: "/img/team-3.jpeg", name: "جسیکا دو" },
+      { img: "/img/team-2.jpeg", name: "رایان تامپسون" },
+      { img: "/img/team-1.jpeg", name: "رومینا حدید" },
+    ],
+  },
+  {
+    img: "/img/home-decor-3.jpeg",
+    title: "مینیمالیست",
+    tag: "پروژه #3",
+    description:
+      "افراد مختلف طعم متفاوتی دارند و انواع مختلفی از موسیقی وجود دارد.",
+    route: "/dashboard/profile",
+    members: [
+      { img: "/img/team-1.jpeg", name: "رومینا حدید" },
+      { img: "/img/team-2.jpeg", name: "رایان تامپسون" },
+      { img: "/img/team-3.jpeg", name: "جسیکا دو" },
+      { img: "/img/team-4.jpeg", name: "الکساندر اسمیت" },
+    ],
+  },
+  {
+    img: "/img/home-decor-4.jpeg",
+    title: "گوتیک",
+    tag: "پروژه #4",
+    description:
+      "چرا کسی باید آبی را بر روی صورتی انتخاب کند؟ صورتی آشکارا رنگ بهتری است.",
+    route: "/dashboard/profile",
+    members: [
+      { img: "/img/team-4.jpeg", name: "الکساندر اسمیت" },
+      { img: "/img/team-3.jpeg", name: "جسیکا دو" },
+      { img: "/img/team-2.jpeg", name: "رایان تامپسون" },
+      { img: "/img/team-1.jpeg", name: "رومینا حدید" },
+    ],
+  },
+];
+
+export default projectsData;