+++ src/pages/dashboard/notifications.jsx
@@ -7,8 +7,12 @@ import {
   CardBody,
 } from "@material-tailwind/react";
 import { InformationCircleIcon } from "@heroicons/react/24/outline";
+import { useLanguage } from "@/context";
 
 export function Notifications() {
+  const { languageData } = useLanguage();
+  const { alertsTitle, alertsDescription, alertsIconTitle } = 
+    languageData.layoutData.alertsPage.notifications;
   const [showAlerts, setShowAlerts] = React.useState({
     blue: true,
     green: true,
@@ -21,7 +25,7 @@ export function Notifications() {
     orange: true,
     red: true,
   });
-  const alerts = ["blue", "green", "orange", "red"];
+  const alerts = ["gray", "green", "orange", "red"];
 
   return (
     <div className="mx-auto my-20 flex max-w-screen-lg flex-col gap-8">
@@ -33,19 +37,16 @@ export function Notifications() {
           className="m-0 p-4"
         >
           <Typography variant="h5" color="blue-gray">
-            Alerts
+            {alertsTitle}
           </Typography>
         </CardHeader>
         <CardBody className="flex flex-col gap-4 p-4">
           {alerts.map((color) => (
             <Alert
               key={color}
-              show={showAlerts[color]}
+              open={showAlerts[color]}
               color={color}
-              dismissible={{
-                onClose: () =>
-                  setShowAlerts((current) => ({ ...current, [color]: false })),
-              }}
+              onClose={() => setShowAlerts((current) => ({ ...current, [color]: false }))}
             >
               A simple {color} alert with an <a href="#">example link</a>. Give
               it a click if you like.
@@ -61,25 +62,22 @@ export function Notifications() {
           className="m-0 p-4"
         >
           <Typography variant="h5" color="blue-gray">
-            Alerts with Icon
+            {alertsIconTitle}
           </Typography>
         </CardHeader>
         <CardBody className="flex flex-col gap-4 p-4">
           {alerts.map((color) => (
             <Alert
               key={color}
-              show={showAlertsWithIcon[color]}
+              open={showAlertsWithIcon[color]}
               color={color}
               icon={
                 <InformationCircleIcon strokeWidth={2} className="h-6 w-6" />
               }
-              dismissible={{
-                onClose: () =>
-                  setShowAlertsWithIcon((current) => ({
-                    ...current,
-                    [color]: false,
-                  })),
-              }}
+              onClose={() => setShowAlertsWithIcon((current) => ({
+                ...current,
+                [color]: false,
+              }))}
             >
               A simple {color} alert with an <a href="#">example link</a>. Give
               it a click if you like.