+++ src/routes.jsx
@@ -2,9 +2,9 @@ import {
   HomeIcon,
   UserCircleIcon,
   TableCellsIcon,
-  BellIcon,
-  ArrowRightOnRectangleIcon,
-  UserPlusIcon,
+  InformationCircleIcon,
+  ServerStackIcon,
+  RectangleStackIcon,
 } from "@heroicons/react/24/solid";
 import { Home, Profile, Tables, Notifications } from "@/pages/dashboard";
 import { SignIn, SignUp } from "@/pages/auth";
@@ -13,7 +13,8 @@ const icon = {
   className: "w-5 h-5 text-inherit",
 };
 
-export const routes = [
+export const routes = {
+  "eng": [
   {
     layout: "dashboard",
     pages: [
@@ -36,9 +37,9 @@ export const routes = [
         element: <Tables />,
       },
       {
-        icon: <BellIcon {...icon} />,
-        name: "notifactions",
-        path: "/notifactions",
+        icon: <InformationCircleIcon {...icon} />,
+        name: "notifications",
+        path: "/notifications",
         element: <Notifications />,
       },
     ],
@@ -48,19 +49,69 @@ export const routes = [
     layout: "auth",
     pages: [
       {
-        icon: <ArrowRightOnRectangleIcon {...icon} />,
+        icon: <ServerStackIcon {...icon} />,
         name: "sign in",
         path: "/sign-in",
         element: <SignIn />,
       },
       {
-        icon: <UserPlusIcon {...icon} />,
+        icon: <RectangleStackIcon {...icon} />,
         name: "sign up",
         path: "/sign-up",
         element: <SignUp />,
       },
     ],
   },
-];
+  ],
+  "fa": [
+    {
+      layout: "dashboard",
+      pages: [
+        {
+          icon: <HomeIcon {...icon} />,
+          name: "داشبورد",
+          path: "/home",
+          element: <Home />,
+        },
+        {
+          icon: <UserCircleIcon {...icon} />,
+          name: "پروفایل",
+          path: "/profile",
+          element: <Profile />,
+        },
+        {
+          icon: <TableCellsIcon {...icon} />,
+          name: "جداول",
+          path: "/tables",
+          element: <Tables />,
+        },
+        {
+          icon: <InformationCircleIcon {...icon} />,
+          name: "اعلانات",
+          path: "/notifications",
+          element: <Notifications />,
+        },
+      ],
+    },
+    {
+      title: "auth pages",
+      layout: "auth",
+      pages: [
+        {
+          icon: <ServerStackIcon {...icon} />,
+          name: "ورود به سیستم",
+          path: "/sign-in",
+          element: <SignIn />,
+        },
+        {
+          icon: <RectangleStackIcon {...icon} />,
+          name: "ثبت نام",
+          path: "/sign-up",
+          element: <SignUp />,
+        },
+      ],
+    },
+  ]
+};
 
 export default routes;