+++ src/widgets/layout/configurator.jsx
@@ -13,6 +13,7 @@ import {
   setSidenavColor,
   setSidenavType,
   setFixedNavbar,
+  useLanguage,
 } from "@/context";
 
 function formatNumber(number, decPlaces) {
@@ -42,13 +43,14 @@ function formatNumber(number, decPlaces) {
 
 export function Configurator() {
   const [controller, dispatch] = useMaterialTailwindController();
+  const { languageData, language, documentDirection, changeLanguage } = useLanguage();
   const { openConfigurator, sidenavColor, sidenavType, fixedNavbar } =
     controller;
   const [stars, setStars] = React.useState(0);
 
   const sidenavColors = {
-    blue: "from-blue-400 to-blue-600",
-    "blue-gray": "from-blue-gray-800 to-blue-gray-900",
+    white: "from-gray-100 to-gray-100 border-gray-200",
+    dark: "from-black to-black border-gray-200",
     green: "from-green-400 to-green-600",
     orange: "from-orange-400 to-orange-600",
     red: "from-red-400 to-red-600",
@@ -63,19 +65,30 @@ export function Configurator() {
       .then((data) => setStars(formatNumber(data.stargazers_count, 1)));
   }, []);
 
+  const setLanguage = React.useCallback((e) => {
+    if(e.target.value) {
+      const docLanguage = e.target.value;
+      changeLanguage(docLanguage);
+    }
+  }, []);
+
   return (
     <aside
-      className={`fixed top-0 right-0 z-50 h-screen w-96 bg-white px-2.5 shadow-lg transition-transform duration-300 ${
-        openConfigurator ? "translate-x-0" : "translate-x-96"
+      className={`fixed overflow-y-auto top-0 right-0 z-50 h-screen w-96 bg-white px-2.5 shadow-lg transition-transform duration-300 ${
+        openConfigurator ? "-translate-x-0" : "translate-x-96"
       }`}
     >
       <div className="flex items-start justify-between px-6 pt-8 pb-6">
         <div>
           <Typography variant="h5" color="blue-gray">
-            Dashboard Configurator
+            { 
+              languageData?.layoutData?.homePage?.dashboradConfiguratorTitle
+            }
           </Typography>
           <Typography className="font-normal text-blue-gray-600">
-            See our dashboard options.
+            { 
+              languageData?.layoutData?.homePage?.dashboradConfiguratorSubtitle
+            }
           </Typography>
         </div>
         <IconButton
@@ -88,8 +101,23 @@ export function Configurator() {
       </div>
       <div className="py-4 px-6">
         <div className="mb-12">
+          <Typography variant="h6" color="blue-gray" className="whitespace-nowrap">
+            { 
+              languageData?.layoutData?.homePage?.dashboradConfiguratorSelectLanguage
+            }
+          </Typography>
+          <div className="w-72">
+            <select defaultValue={localStorage.getItem("lang") || "eng"} onClick={setLanguage}>
+              <option value="eng">English</option>
+              <option value="fa">فارسی</option>
+            </select>
+          </div>
+        </div>
+        <div className="mb-12">
           <Typography variant="h6" color="blue-gray">
-            Sidenav Colors
+            { 
+              languageData?.layoutData?.homePage?.dashboradSidenavColorsTitle
+            }
           </Typography>
           <div className="mt-3 flex items-center gap-2">
             {Object.keys(sidenavColors).map((color) => (
@@ -107,10 +135,14 @@ export function Configurator() {
         </div>
         <div className="mb-12">
           <Typography variant="h6" color="blue-gray">
-            Sidenav Types
+           { 
+              languageData?.layoutData?.homePage?.dashboradSidenavTypesTitle
+            }
           </Typography>
           <Typography variant="small" color="gray">
-            Choose between 3 different sidenav types.
+           { 
+              languageData?.layoutData?.homePage?.dashboradSidenavTypesSubtitle
+            }
           </Typography>
           <div className="mt-3 flex items-center gap-2">
             <Button
@@ -137,7 +169,7 @@ export function Configurator() {
           <hr />
           <div className="flex items-center justify-between py-5">
             <Typography variant="h6" color="blue-gray">
-              Navbar Fixed
+              {languageData?.layoutData?.homePage?.dashboradSidenavNavbarType}
             </Typography>
             <Switch
               id="navbar-fixed"
@@ -213,7 +245,6 @@ export function Configurator() {
           <div className="mt-4 flex justify-center gap-2">
             <Button
               variant="gradient"
-              color="blue-gray"
               className="flex items-center gap-2"
             >
               <i className="fa-brands fa-twitter text-white" />
@@ -221,7 +252,6 @@ export function Configurator() {
             </Button>
             <Button
               variant="gradient"
-              color="blue-gray"
               className="flex items-center gap-2"
             >
               <i className="fa-brands fa-facebook text-white" />