+++ tailwind.config.cjs
@@ -6,5 +6,5 @@ module.exports = withMT({
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require('tailwindcss-rtl')],
 });