+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [2.1.0] 2023-11-14
+- Upgrade to Material Tailwind `v2`
+
+## [2.0.1] 2023-03-29
+
+- Add nepcha scripts
+
 ## [2.0.0] 2022-11-02
 
 - Update the entire structure